separating logic for the alphas to their own module
